condense attribute pointer cache functions and move handleSpecialState8_clearAsDepth to common renderer
